MAINT: move axis_keepdims_wrapper to decorators
